Interaction Event System (#120)
## Purpose
This system adds components to trigger interactions between anything in the game. It uses an event based architecture with a very minimal implementation. It supports dependencies between events, such that some events may block others.

## Changes to existing Unity files
- `Assets/Prefab/Furniture/Machinery/locker.prefab`: Added new interactions
- `Assets/Prefab/Human/Human.prefab`: Added interactor that actually sends events
- `Assets/Prefab/Items/Functional/Tools/crowbar.prefab`: Added new interactions

## Technical implementation notes
The core of the system is the `IInteraction` interface that defines the methods needed to receive events, and `InteractionReceiver` which actually receives the events and passes them on to the interactions.
An abstract class simply named `Interaction` inherits from `ScriptableObject` and can be added on to `InteractionReceiver` components to easily add simple interactions to an object. This is mostly intended for fully or mostly stateless interactions, as they aren't attached to the object itself.
For stateful interactions (such as opening a door), it is recommended to extend from `MonoBehaviour`, implement the `IInteraction` interface and require a `InteractionReceiver` sibling component.
It is however possible to implement the interface from anything and there is no requirements that events actually originate from a GameObject. An example extension of this system would be to have a developer console, where one can input events directly to debug and simulate certain situations or to drive admin commands. Any event can be sent to anything from anywhere, so a door doesn't need to care if it is opened by a player, a signal, or an admin command.

## Known issues
- Not everything uses this new system yet

Closes #87