Make Internal Clothing UI able to be toggled by clicks and pushing `. Keep selected hand highlighted on hover. (#692)
* Make Internal Clothing UI collapsible by clicks.

The Internal Clothing UI Can now be collapsed by clicking on a button in the down left part of the screen. See issue #661 .

* Make Internal Clothing UI collapsible by pressing ` . Renamed the concerned files..

See issue #692.  The Clothing UI can now be collapsed and expanded by pressing `. Renamed ClickExpandInternalClothing into something reflecting better the behavior of the script, as it does not only changes the UI on click but also on pressing `.

* Update HandsUI.cs. Keep the selected hand highlighted when hovered.

See issue #691 .

* Update InventoryUIBeep.prefab

Added an ItemDisplay behind the toggling symbol in the down left corner.

* Update InventoryUIBeep.prefab

Modified the color of the arrow on the expand picture so it matches better the other elements of the UI.