Examine System: hover text implementation and text visibility (#777)
* CompositeItemSelector.cs no longer records the mesh texture.

Upon review of the code, it was clear that the recorded texture was never actually used anywhere. Additionally, by allocating the texture to the singleColourMaterial, it was changing that file every time the Examine system was used. This caused the file to want to update within version control after every commit.

* Created a hover text field on the Examine UI.

This field is configured to display the name of the object under the mouse cursor, even when the examine button (shift) is not held down. To accommodate this, all IExamineData are now required to have a GetName() method.

* Examine now correctly identifies SkinnedMeshRenderers.

Previously, the SkinnedMeshRenderer mesh used by the Examine system was simply the model bind pose, which gave incorrect examine results. Now, the mesh is baked prior to rendering, ensuring that the Examine system returns a pixel-perfect match.

* Added an outline to the Examine text.

* Added a new material asset for Examine text font.

The new asset is required so that the Examine text can have an outline, while the original font can be used elsewhere (i.e. Chat system) without an outline.

* Added the new examine font to the simple text and simple image UI's.

* Turned off hover text when another examine UI element is in use.

Allowed setting in ExamineUI to turn off hover text completely.

* Removed unnecessary matrix recalculation calls.

* Examine font sizes changed.
- Hover text: 30pt
- Simple text (name): 30pt
- Simple text (description): 26pt
- Simple image: 30pt

* Changed simple examine description to bold text.

* Centred hover text. Removed outline from chat font.