New Container Editor, multiple container on a single game object (#798)
Simplifies adding a container to a game object, as the current system is over complicated for contributors. All relevant container parameters, such as it's type, it's size, if it displays objects, and more, are in a single script, containerDescriptor.
This script has a custom editor that allows contributors to build a container in a coherent manner.