Networking: Update Mirror to v49.0.1 (#814)
* Direct replacement of Mirror folder with v49.0.1. Not yet functional.
(Test and Examples folders were excluded)

* Updates ignoreAuthority to requiresAuthority.

* Updated MessageBase to NetworkMessage

* Updated ClientScene to NetworkClient

* Updated various obsolete methods to ReadUInt and WriteUInt.

* Removed calls to obsolete Logger.

* NetworkMessages converted to structs

* Adds the currentSender accessor to Mirror's RemoteCallHelper.

* Added connectionFailed Action back onto Telepathy/Client.cs

* Changed API compatibility level to .Net 4.x

* Removed null-coalescing operator from a NetworkMessage instance
NetworkMessages are now structs rather than classes, so cannot be null.

* NetworkConnection argument removed from NetworkClient.RegisterHandler() call.

* Ensure NetworkClient is ready before calling AddPlayer.

* Update Project Settings

* Removed subordinate NetworkIdentities and NetworkTransforms from prefab children.
Also removed 'Missing' script from TankStorageUnit.prefab.

* Commented out SubstanceDisplay.RpcUpdateDisplay()

This method suffered from a NullReferenceException in the message handler. The latest version of Mirror will automatically disconnect clients who have exceptions in NetworkMessages.

* Updated LoginNetworkManager to more clearly define Client and Server methods.

SpawnPlayerAfterRoundStart() is a server-only method. Previously, it called SpawnPlayerWithoutLoginServer(), which should be a client-only method because it sends a CharacterSelectMessage. CharacterSelectMessages should only be sent from client to server. Because it was being sent from server to client instead, it was causing errors as there was not a registered handler of these messages on the client.

* Removed a missing script from Human.

* Removed NetworkIdentities from Human prefab's hands.

* Removed subordinate NetworkIdentities / NetworkTransforms:
- Vent.prefab
- SodaVendor.prefab
- Fridge.prefab

* Removed subordinate NetworkIdentities from CellCharger.prefab.

* Essential scene changes to MiniStation.unity.

Reloaded MiniStation tilemap from save. This is to remove any tile objects that were inadvertently added outside the Tilemap Editor.

Removed Janicart. The Janicart prefab is broken because of subordinate NetworkIdentities, and will be addressed in future work.

* Put Burger.prefab back into the correct (item) layer.

* Added NetworkStartPositions to MiniStation.

A somewhat disturbing bug had emerged where the Human prefab looked monstrous on the server (though looked normal on the client). It appeared that this was caused by the players instantiating in exactly the same position, as it did not occur when the players had moved away prior to the next one embarking. The four NetworkStartPositions added ensures that players will not routinely spawn as monsters.

It is possible that players *might* still look monstrous if they move onto another spawn point as a new player spawns there, but I think that is unlikely enough not to worry about for the moment.