Fix issues with basic Substance system (#823)
* Basic scene and prefab changes to support substances.

Added Registry, Water Cooler and Booze Dispenser to scene.

Added Kahlua back to the Substance Registry.

* Fixed issues in SubstanceContainer.

Spelling corrected on CanTransfer() function.

Previous PR #499 had added the CanTransfer() function. However, the boolean returned from this function was incorrect. You should be able to transfer if the container is **not** locked, rather than if it is.

* Fixes to SubstanceDispensers.

SubstanceDispenserEditor has been temporarily disabled due to incompatible changes to SubstanceDispenser.

SubstanceDispenser can now dispense multiple substance from the same script, and will display their name in the interaction.

SubstanceDispensers now can be configured to dispense either milliliters or moles. Bartenders (for example) would expect Booze Dispensers to dispense in mL, but Chemists would need moles for accurate chemical reactions. TODO: Fix BlackRussian recipe proportions to account for mL.

* Gave the beakers their mesh colliders back.

* Scene changes to add some beakers.

* Amended SubstanceDisplay to allow viewing substances on client.

* Adjusted recipe for BlackRussian cocktail

It is now 3 parts vodka, 2 parts kahlua by volume, rather than by moles.

* Re-created Substance Registry prefab.

Substance Registry has been added to the Managers prefab, which is held in the ServerLobby scene. This will allow the Substance Registry to be loaded onto any map, not just MiniStation. All future substances and recipes should be loaded onto the Substance Registry prefab.

* Removed a few tile objects incorrectly added as items.
- 1 x Fridge
- 1 x Disposal Bin
- 1 x Locker

* Removed subordinate NetworkIdentities / NetworkTransforms from Water Cooler

* Tilemap cleanup.

* Deleted old Registry prefab, and moved Substance Registry prefab to Content.

* Make sure SubstanceDisplay is initialized before accepting RPC's.

* Prevented DispenseSubstanceInteraction from occurring when receptacle is already full.

* WaterCooler now dispenses 300mL instead of 250mL, so it fills paper cup in one interaction.

* Synchronised the volume and remainingVolume variables of SubstanceContainer from server to client.

Because remainingVolume is required for the CanInteract function of the DispenseSubstanceInteraction, it needed to be synchronised. Otherwise, the CanInteract check would pass on the client even when the container is full (so all the Dispense icons would show up) but it would fail on the server.

* Server tags added to relevant methods, and minor fixes.